requirements.txt
streamlit
requests
python-dotenv
regex